* Refactor application registry tests. Remove extra fake FreameworkUtil. * Exclude osgi classes from tests execution,
