+++ flow-server/src/main/java/com/vaadin/flow/server/osgi/OSGiAccess.java
@@ -37,13 +37,13 @@ import javax.servlet.ServletException;
 import javax.servlet.ServletRegistration;
 import javax.servlet.annotation.HandlesTypes;
 
+import com.vaadin.flow.internal.UsageStatistics;
 import org.osgi.framework.Bundle;
 import org.slf4j.LoggerFactory;
 
 import com.googlecode.gentyref.GenericTypeReflector;
 import com.vaadin.flow.internal.AnnotationReader;
 import com.vaadin.flow.internal.ReflectTools;
-import com.vaadin.flow.internal.UsageStatistics;
 import com.vaadin.flow.router.HasErrorParameter;
 
 import net.bytebuddy.ByteBuddy;
@@ -267,6 +267,7 @@ public final class OSGiAccess {
                 Class.forName("org.osgi.framework.FrameworkUtil");
 
                 UsageStatistics.markAsUsed("flow/osgi", getOSGiVersion());
+
                 return true;
             } catch (ClassNotFoundException exception) {
                 return false;
@@ -276,20 +277,15 @@ public final class OSGiAccess {
         /**
          * Tries to detect the version of the OSGi framework used.
          *
-         * @return the used OSGi version or {@code null} if not able to detect
-         *         it
+         * @return the used OSGi version or {@code null} if not able to detect it
          */
         private static String getOSGiVersion() {
             try {
-                Bundle osgiBundle = org.osgi.framework.FrameworkUtil
-                        .getBundle(Bundle.class);
+                Bundle osgiBundle = org.osgi.framework.FrameworkUtil.getBundle(Bundle.class);
                 return osgiBundle.getVersion().toString();
             } catch (Throwable throwable) {
-                // just eat it so that any failure in the version detection
-                // doesn't break OSGi usage
-                LoggerFactory.getLogger(OSGiAccess.class).info(
-                        "Unable to detect used OSGi framework version due to "
-                                + throwable.getMessage());
+                // just eat it so that any failure in the version detection doesn't break OSGi usage
+                LoggerFactory.getLogger(OSGiAccess.class).info("Unable to detect used OSGi framework version due to " + throwable.getMessage());
             }
             return null;
         }
