+++ flow-server/src/test/java/com/vaadin/flow/server/startup/ServletDeployerTest.java
@@ -1,5 +1,34 @@
 package com.vaadin.flow.server.startup;
 
+import javax.servlet.Registration;
+import javax.servlet.Servlet;
+import javax.servlet.ServletConfig;
+import javax.servlet.ServletContext;
+import javax.servlet.ServletContextEvent;
+import javax.servlet.ServletRegistration;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import org.easymock.Capture;
+import org.easymock.CaptureType;
+import org.easymock.EasyMock;
+import org.junit.Before;
+import org.junit.Test;
+
+import com.vaadin.flow.component.Component;
+import com.vaadin.flow.router.Route;
+import com.vaadin.flow.router.internal.RouteUtil;
+import com.vaadin.flow.server.Constants;
+import com.vaadin.flow.server.RouteRegistry;
+import com.vaadin.flow.server.VaadinServlet;
+
 import static java.util.Collections.emptyList;
 import static java.util.Collections.emptyMap;
 import static java.util.Collections.singletonList;
@@ -14,37 +43,6 @@ import static org.easymock.EasyMock.newCapture;
 import static org.easymock.EasyMock.replay;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
-
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import javax.servlet.Registration;
-import javax.servlet.Servlet;
-import javax.servlet.ServletConfig;
-import javax.servlet.ServletContext;
-import javax.servlet.ServletContextEvent;
-import javax.servlet.ServletRegistration;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-
-import org.easymock.Capture;
-import org.easymock.CaptureType;
-import org.easymock.EasyMock;
-import org.junit.Before;
-import org.junit.Test;
-
-import com.vaadin.flow.component.Component;
-import com.vaadin.flow.router.Route;
-import com.vaadin.flow.router.internal.RouteUtil;
-import com.vaadin.flow.server.Constants;
-import com.vaadin.flow.server.RouteRegistry;
-import com.vaadin.flow.server.VaadinServlet;
-import com.vaadin.flow.server.osgi.OSGiAccess;
 
 public class ServletDeployerTest {
     private final ServletDeployer deployer = new ServletDeployer();
@@ -89,13 +87,6 @@ public class ServletDeployerTest {
     public void clearCaptures() {
         servletNames.reset();
         servletMappings.reset();
-
-        if (OSGiAccess.getInstance().getOsgiServletContext() != null) {
-            ApplicationRouteRegistry
-                    .getInstance(
-                            OSGiAccess.getInstance().getOsgiServletContext())
-                    .clean();
-        }
     }
 
     @Test
@@ -261,9 +252,8 @@ public class ServletDeployerTest {
             expect(contextMock.getAttribute(RouteRegistry.class.getName()))
                     .andAnswer(() -> {
                         ApplicationRouteRegistry registry = new ApplicationRouteRegistry();
-                        RouteUtil.setNavigationTargets(
-                                Collections.singleton(ComponentWithRoute.class),
-                                registry);
+                        RouteUtil.setNavigationTargets(Collections
+                                .singleton(ComponentWithRoute.class), registry);
                         return registry;
                     }).anyTimes();
         } else {
@@ -294,8 +284,7 @@ public class ServletDeployerTest {
         Capture<String> parameterNameCapture = newCapture();
         expect(registrationMock.getInitParameter(capture(parameterNameCapture)))
                 .andAnswer(() -> initParameters
-                        .get(parameterNameCapture.getValue()))
-                .anyTimes();
+                        .get(parameterNameCapture.getValue())).anyTimes();
         replay(registrationMock);
         return registrationMock;
     }
